third_party/blink/web_tests/wpt_internal/attribution-reporting/aggregatable-report-deduplication.sub.https.html

<!doctype html>
<meta charset=utf-8>
<meta name=timeout content=long>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/helpers.js"></script>
<script>
const aggregatable_dedup_promise_test = (testCase) => {
  attribution_reporting_promise_test(async t => {
    const host = 'https://{{host}}';

    const source = {
      source_event_id: generateSourceEventId(),
      destination: host,
      aggregation_keys: {
        campaignCounts: '0x159',
      },
      debug_key: '456',
      debug_reporting: true,
      filter_data: {
        'x': ['y', 'z'],
      }
    };
    await registerAttributionSrc({
      source,
      method: 'open',
      cookie: attributionDebugCookie,
    });
    await waitForSourceToBeRegistered(source.source_event_id);

    const trigger = {
      aggregatable_trigger_data: [{
        key_piece: '0x400',
        source_keys: ['campaignCounts'],
      }],
      aggregatable_values: {
        campaignCounts: 32768,
      },
      aggregatable_deduplication_keys: [{
        'deduplication_key': '666',
      }],
      debug_reporting: true,
    };

    await registerAttributionSrc({
      trigger,
    });

    const payload = await pollAggregatableReports();
    assert_equals(payload.reports.length, 1);
    const report = JSON.parse(payload.reports[0].body);
    assert_own_property(report, 'shared_info');
    const shared_info = JSON.parse(report.shared_info);
    assert_equals(shared_info.attribution_destination, host);

    trigger.aggregatable_deduplication_keys =
        testCase.aggregatable_deduplication_keys;

    await registerAttributionSrc({
      trigger,
      cookie: attributionDebugCookie,
    });

    if (testCase.expectDeduplicated) {
      const debugPayload = await pollVerboseDebugReports();
      assert_equals(debugPayload.reports.length, 1);
      const debugReport = JSON.parse(debugPayload.reports[0].body);
      assert_equals(debugReport.length, 1);
      assert_equals(debugReport[0].type, 'trigger-aggregate-deduplicated');
      assert_own_property(debugReport[0], 'body');
      const debugReportBody = debugReport[0].body;
      assert_equals(debugReportBody.attribution_destination, host);
      assert_equals(debugReportBody.source_event_id, source.source_event_id);
      assert_equals(debugReportBody.source_site, host);
      assert_equals(debugReportBody.source_debug_key, '456');
      assert_not_own_property(debugReportBody, 'trigger_debug_key');
      return;
    }

    const payload2 = await pollAggregatableReports();
    assert_equals(payload2.reports.length, 1);
  }, testCase.testName);
};

[
  {
    testName: 'No deduplication key',
    aggregatable_deduplication_keys: [],
    expectDeduplicated: false,
  },
  {
    testName: 'Filters match',
    aggregatable_deduplication_keys: [
      {
        'deduplication_key': '666',
        'filters': {'x': ['y']},
      }
    ],
    expectDeduplicated: true,
  },
  {
    testName: 'Filters does not match',
    aggregatable_deduplication_keys: [
      {
        'deduplication_key': '666',
        'filters': {'x': ['x']},
      }
    ],
    expectDeduplicated: false,
  },
  {
    testName: 'Negated filters match',
    aggregatable_deduplication_keys: [
      {
        'deduplication_key': '666',
        'not_filters': {'x': ['y']},
      }
    ],
    expectDeduplicated: false,
  },
  {
    testName: 'Negated filters does not match',
    aggregatable_deduplication_keys: [
      {
        'deduplication_key': '666',
        'not_filters': {'x': ['x']},
      }
    ],
    expectDeduplicated: true,
  },
  {
    testName: 'Different deduplication key',
    aggregatable_deduplication_keys: [
      {
        'deduplication_key': '456',
        'filters': {'x': ['y']}
      }
    ],
    expectDeduplicated: false,
  },
  {
    testName: 'Null deduplication key',
    aggregatable_deduplication_keys: [
      {
        'filters': {'x': ['y']}
      }
    ],
    expectDeduplicated: false,
  },
  {
    testName: 'Multiple deduplication keys',
    aggregatable_deduplication_keys: [
      {
        'deduplication_key': '555',
        'filters': {'x': ['x']},
      },
      {
        'deduplication_key': '666',
        'filters': {'x': ['y']},
      },
      {
        'deduplication_key': '777',
        'filters': {'x': ['z']},
      },
    ],
    expectDeduplicated: true,
  },
].forEach(aggregatable_dedup_promise_test);

</script>
